requirements.txt
# Core
httpx>=0.27.0
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Optimization
pulp>=2.8.0

# LLM Integration (optional - comment out if not using)
# litellm>=1.30.0

# CLI
typer[all]>=0.12.0
rich>=13.7.0

# Web UI
streamlit>=1.31.0

# Database
sqlite-utils>=3.36.0
